#86: refactor examples into separate components
