+++ src/app/app.module.ts
@@ -3,10 +3,6 @@ import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-
-
-import { NgxMatSelectSearchModule } from './mat-select-search/ngx-mat-select-search.module';
-
 import {
   MatButtonModule,
   MatFormFieldModule,
@@ -14,7 +10,15 @@ import {
   MatSelectModule,
   MatToolbarModule
 } from '@angular/material';
-import {AppComponent} from './app.component';
+
+import { NgxMatSelectSearchModule } from './mat-select-search/ngx-mat-select-search.module';
+
+import { AppComponent } from './app.component';
+import { SingleSelectionExampleComponent } from './examples/01-single-selection-example/single-selection-example.component';
+import { MultipleSelectionExampleComponent } from './examples/02-multiple-selection-example/multiple-selection-example.component';
+import { CustomClearIconExampleComponent } from './examples/03-custom-clear-icon-example/custom-clear-icon-example.component';
+import { OptionGroupsExampleComponent } from './examples/04-option-groups-example/option-groups-example.component';
+import { ServerSideSearchExampleComponent } from './examples/05-server-side-search-example/server-side-search-example.component';
 
 /**
  * NgModule that includes all Material modules that are required to serve the app.
@@ -40,7 +44,14 @@ export class MaterialModule {}
     BrowserAnimationsModule,
     NgxMatSelectSearchModule
   ],
-  declarations: [AppComponent],
+  declarations: [
+    AppComponent,
+    SingleSelectionExampleComponent,
+    MultipleSelectionExampleComponent,
+    CustomClearIconExampleComponent,
+    OptionGroupsExampleComponent,
+    ServerSideSearchExampleComponent
+  ],
   bootstrap: [AppComponent],
   providers: []
 })
