+++ src/app/examples/01-single-selection-example/single-selection-example.component.html
@@ -1,0 +1,14 @@
+<h3>Single selection</h3>
+<p>
+  <mat-form-field>
+    <mat-select [formControl]="bankCtrl" placeholder="Bank" #singleSelect>
+      <ngx-mat-select-search [formControl]="bankFilterCtrl"></ngx-mat-select-search>
+      <mat-option *ngFor="let bank of filteredBanks | async" [value]="bank">
+        {{bank.name}}
+      </mat-option>
+    </mat-select>
+  </mat-form-field>
+</p>
+<p>
+  Selected Bank: {{bankCtrl.value?.name}}
+</p>
