+++ src/app/examples/02-multiple-selection-example/multiple-selection-example.component.html
@@ -1,0 +1,18 @@
+
+<h3>Multiple selection</h3>
+<p>
+  <mat-form-field>
+    <mat-select [formControl]="bankMultiCtrl" placeholder="Banks" [multiple]="true" #multiSelect>
+      <ngx-mat-select-search [formControl]="bankMultiFilterCtrl"></ngx-mat-select-search>
+      <mat-option *ngFor="let bank of filteredBanksMulti | async" [value]="bank">
+        {{bank.name}}
+      </mat-option>
+    </mat-select>
+  </mat-form-field>
+</p>
+<p>
+  Selected Banks:
+</p>
+<ul *ngFor="let bank of bankMultiCtrl?.value">
+  <li>{{bank.name}}</li>
+</ul>
