+++ src/app/examples/04-option-groups-example/option-groups-example.component.ts
@@ -1,0 +1,87 @@
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { ReplaySubject, Subject } from 'rxjs';
+import { FormControl } from '@angular/forms';
+import { takeUntil } from 'rxjs/operators';
+
+import { BankGroup, BANKGROUPS } from '../demo-data';
+
+
+@Component({
+  selector: 'app-option-groups-example',
+  templateUrl: './option-groups-example.component.html',
+  styleUrls: ['./option-groups-example.component.scss']
+})
+export class OptionGroupsExampleComponent implements OnInit, OnDestroy {
+
+  /** list of bank groups */
+  protected bankGroups: BankGroup[] = BANKGROUPS;
+
+  /** control for the selected bank for option groups */
+  public bankGroupsCtrl: FormControl = new FormControl();
+
+  /** control for the MatSelect filter keyword for option groups */
+  public bankGroupsFilterCtrl: FormControl = new FormControl();
+
+  /** list of bank groups filtered by search keyword for option groups */
+  public filteredBankGroups: ReplaySubject<BankGroup[]> = new ReplaySubject<BankGroup[]>(1);
+
+  /** Subject that emits when the component has been destroyed. */
+  protected _onDestroy = new Subject<void>();
+
+
+  constructor() { }
+
+  ngOnInit() {
+    // load the initial bank list
+    this.filteredBankGroups.next(this.copyBankGroups(this.bankGroups));
+
+    // listen for search field value changes
+    this.bankGroupsFilterCtrl.valueChanges
+      .pipe(takeUntil(this._onDestroy))
+      .subscribe(() => {
+        this.filterBankGroups();
+      });
+  }
+
+  ngOnDestroy() {
+    this._onDestroy.next();
+    this._onDestroy.complete();
+  }
+
+
+  protected filterBankGroups() {
+    if (!this.bankGroups) {
+      return;
+    }
+    // get the search keyword
+    let search = this.bankGroupsFilterCtrl.value;
+    const bankGroupsCopy = this.copyBankGroups(this.bankGroups);
+    if (!search) {
+      this.filteredBankGroups.next(bankGroupsCopy);
+      return;
+    } else {
+      search = search.toLowerCase();
+    }
+    // filter the banks
+    this.filteredBankGroups.next(
+      bankGroupsCopy.filter(bankGroup => {
+        const showBankGroup = bankGroup.name.toLowerCase().indexOf(search) > -1;
+        if (!showBankGroup) {
+          bankGroup.banks = bankGroup.banks.filter(bank => bank.name.toLowerCase().indexOf(search) > -1);
+        }
+        return bankGroup.banks.length > 0;
+      })
+    );
+  }
+
+  protected copyBankGroups(bankGroups: BankGroup[]) {
+    const bankGroupsCopy = [];
+    bankGroups.forEach(bankGroup => {
+      bankGroupsCopy.push({
+        name: bankGroup.name,
+        banks: bankGroup.banks.slice()
+      });
+    });
+    return bankGroupsCopy;
+  }
+}
