+++ src/app/mat-select-search/ngx-mat-select-search.module.ts
@@ -11,6 +11,9 @@ import { MatButtonModule, MatInputModule, MatIconModule } from '@angular/materia
 import { CommonModule } from '@angular/common';
 import { MatSelectSearchClearDirective } from './mat-select-search-clear.directive';
 
+export const MatSelectSearchVersion = '1.4.3';
+
+
 @NgModule({
   imports: [
     CommonModule,
